Extract shortcut formatter to separate file
