--- shortcuts-raycast-extension/src/view/base-key-override.ts
@@ -1,20 +1,0 @@
-export const baseKeySymbolOverride: Map<string, string> = new Map([
-  ["left", "←"],
-  ["right", "→"],
-  ["up", "↑"],
-  ["down", "↓"],
-  ["pageup", "PgUp"],
-  ["pagedown", "PgDown"],
-  ["home", "Home"],
-  ["end", "End"],
-  ["space", "Space"],
-  ["capslock", "⇪"],
-  ["backspace", "⌫"],
-  ["tab", "⇥"],
-  ["esc", "⎋"],
-  ["enter", "↩"],
-  ["cmd", "⌘"],
-  ["ctrl", "⌃"],
-  ["opt", "⌥"],
-  ["shift", "⇧"],
-]);
