+++ shortcuts-raycast-extension/src/view/hotkey-text-formatter.ts
@@ -1,0 +1,39 @@
+import { SectionShortcut } from "../model/internal/internal-models";
+import { modifierSymbols } from "../model/internal/modifiers";
+
+const baseKeySymbolOverride: Map<string, string> = new Map([
+  ["left", "←"],
+  ["right", "→"],
+  ["up", "↑"],
+  ["down", "↓"],
+  ["pageup", "PgUp"],
+  ["pagedown", "PgDown"],
+  ["home", "Home"],
+  ["end", "End"],
+  ["space", "Space"],
+  ["capslock", "⇪"],
+  ["backspace", "⌫"],
+  ["tab", "⇥"],
+  ["esc", "⎋"],
+  ["enter", "↩"],
+  ["cmd", "⌘"],
+  ["ctrl", "⌃"],
+  ["opt", "⌥"],
+  ["shift", "⇧"],
+]);
+
+function overrideSymbolIfPossible(base: string) {
+  if (baseKeySymbolOverride.has(base)) {
+    return baseKeySymbolOverride.get(base);
+  }
+  return base.toUpperCase();
+}
+
+export function generateHotkeyText(shortcut: SectionShortcut): string {
+  return shortcut.sequence
+    .map((atomicShortcut) => {
+      const modifiersText = atomicShortcut.modifiers.map((modifier) => modifierSymbols.get(modifier)).join("") ?? "";
+      return modifiersText + overrideSymbolIfPossible(atomicShortcut.base);
+    })
+    .join(" ");
+}
